Add response and status code for invalid id and posts not found

+++ controllers/post.js
@@ -1,3 +1,4 @@
+const { default: mongoose } = require('mongoose');
 const Post = require('../models/post');
 const { check, validationResult } = require('express-validator');
 
@@ -23,7 +24,7 @@ exports.create_post = [
 
       res.status(200).json(post);
     } catch (error) {
-      res.status(500).send(error);
+      res.status(500).json({ error: 'Server error, post not saved' });
     }
   }
 ]
@@ -35,20 +36,31 @@ exports.update_post = [
   async (req, res) => {
     try {
       if (!errors.isEmpty()) {
-        res.status(422).json({ errors: errors.array() });
-      } else {
-        const updatedPost = new Post({
-          title: req.body.title,
-          text: req.body.text,
-          author: req.body.author,
-          timestamp: req.body.timestamp,
-          _id: req.params.id,
-        });
-      
-        const post = await Post.findByIdAndUpdate(req.params.id, updatedPost, {});
-      
-        res.status(200).json(post);
+        return res.status(422).json({ errors: errors.array() });
       }
+
+      const _id = req.params.id;
+      const isValidId = mongoose.Types.ObjectId.isValid(_id);
+
+      if (!isValidId) {
+        return res.status(422).json({ error: 'Id is not valid' });
+      }
+
+      const updatedPost = new Post({
+        title: req.body.title,
+        text: req.body.text,
+        author: req.body.author,
+        timestamp: req.body.timestamp,
+        _id: req.params.id,
+      });
+      
+      const post = await Post.findByIdAndUpdate(req.params.id, updatedPost, {});
+
+      if (post === null) {
+        return res.status(404).json({ error: 'Post not found'});
+      }
+
+      res.status(200).json(post);
     } catch (error) {
       res.status(500).send(error);
     }
@@ -58,16 +70,35 @@ exports.update_post = [
 exports.get_posts = async (req, res) => {
   const posts = await Post.find({});
 
+  if (!posts.length) {
+    return res.status(404).json({ error: 'Posts not found' });
+  }
+
   res.status(200).json(posts);
 };
 
 exports.get_post = async (req, res) => {
+  const _id = req.params.id;
+  const isValidId = mongoose.Types.ObjectId.isValid(_id);
+  if (!isValidId) {
+    return res.status(422).json({ error: 'The id is invalid' });
+  }
   const post = await Post.findById(req.params.id);
 
+  if (post === null) {
+    return res.status(404).json({ error: 'Post not found'})
+  }
   res.status(200).json(post);
 };
 
 exports.delete_post = async (req, res) => {
+  const _id = req.params.id;
+  const isValidId = mongoose.Types.ObjectId.isValid(_id);
+
+  if (!isValidId) {
+    return res.status(422).json({ error: 'The is is invalid' });
+  }
+
   const post = await Post.findByIdAndDelete(req.params.id);
 
   res.status(200).json(post);
